styling, readme, comments

+++ client/src/pages/Saved.js
@@ -3,33 +3,32 @@ import Navbar from "../components/Navbar";
 import Jumbotron from "../components/Jumbotron";
 import API from "../utils/API";
 import { Container } from "../components/Grid";
-import { List, ListItem } from "../components/List";
-import { FormBtn } from "../components/Form";
+import Card from "../components/Card";
 
 class Saved extends Component {
-
+    // Set State
     state = {
         books: []
-    }
-
+    };
+    // When page loads, get saved books from db
     componentDidMount() {
         this.loadBooks()
-    }
-
+    };
+    // Method for loading books from db
     loadBooks = () => {
         API.loadSavedBooks()
             .then(res => this.setState({ books: res.data }))
             .catch(err => console.log(err));
-    }
-
+    };
+    // Method for deleting a saved book, asking for a confirm to make sure
     handleDelete = (id) => {
-        console.log(id);
-        API.deleteBook(id)
-        .then(res => this.loadBooks())
-        .catch(err => console.log(err));
-    }
-
-
+        if (window.confirm("Delete this book from your saved books?")) {
+            API.deleteBook(id)
+                .then(res => this.loadBooks())
+                .catch(err => console.log(err));
+        };
+    };
+    // Rendering, mapping over saved books to render each
     render() {
         return (
             <div>
@@ -39,18 +38,17 @@ class Saved extends Component {
                         <h1>Saved Books</h1>
                         <p>A list of your current saved books...</p>
                     </Jumbotron>
-                    <List>
                     {this.state.books.map(book => (
-                        <ListItem key={book._id}>
-                            <h3>{book.title}</h3>
-                            <h4>{book.authors}</h4>
-                            <p>{book.description}</p>
-                            <a href={book.link}>Info Link</a>
-                            <img src={book.image} alt={book.title} />
-                            <FormBtn onClick={() => this.handleDelete(book._id)}>Delete Book</FormBtn>
-                        </ListItem>
+                        <Card key={book._id}
+                            title={book.title}
+                            authors={book.authors}
+                            description={book.description}
+                            image={book.image}
+                            infoLink={book.link}
+                            display="Delete Book"
+                            FormBtn onClick={() => this.handleDelete(book._id)}
+                        />
                     ))}
-                    </List>
                 </Container>
             </div>
         )
